ESDEV-3589 Delete table only if it exists
AddoDBlite was ignoring that DELETE table returns error when table does not exist.
Doctrine throws exception which makes some random failures on test run.

From now on:
If table exist, but it's data is wrong:
- delete table
- restore original table
If table does not exist:
- restore original table

+++ library/Services/Library/DatabaseRestorer/DatabaseRestorerToFile.php
@@ -133,7 +133,9 @@ class DatabaseRestorerToFile implements DatabaseRestorerInterface
         }
 
         $database = oxDb::getDb();
-        $database->execute("DROP TABLE `$table`");
+        if ($database->getOne("SHOW TABLES LIKE '$table'")) {
+            $database->execute("DROP TABLE `$table`");
+        }
 
         $databaseHandler = $this->getDatabaseHandler();
         $directory = $this->getDumpDirectory();
